feat: add problem page layout(#75)
