+++ src/components/ui/atoms/Button/button.module.scss
@@ -1,5 +1,5 @@
 .button {
-  border-radius: 24px;
+  border-radius: 30px;
   border: none;
   background-color: white;
 
@@ -8,11 +8,11 @@
   }
 
   &.medium {
-    padding: 15px 36px;
+    padding: 14px 36px;
   }
 
   &.large {
-    padding: 20px 76px;
+    padding: 14px 92px;
   }
 
   /*
@@ -34,7 +34,6 @@
   &.answer {
     border: 1px solid hsl(0, 0%, 75%);
     background-color: #EBEBEB;
-    width: 80vw;
     font-size: 16px;
     font-weight: 600;
     &:hover {
