+++ src/components/ui/organism/SelectAnswer/index.tsx
@@ -5,12 +5,12 @@ type Props = React.HTMLAttributes<HTMLButtonElement> & {
   array: string[];
 };
 
-export default function SelectAnswer({ array }: Props) {
+export default function SelectAnswer({ onClick, array }: Props) {
   return (
     <div className={$.selectAnswer}>
       {array.map((item, index) => {
         return (
-          <Button key={index} onClick={() => ""} size="medium" variant="answer">
+          <Button key={index} onClick={onClick} size="large" variant="answer">
             {item}
           </Button>
         );
