+++ src/components/ui/template/ProblemLayout/index.module.scss
@@ -1,0 +1,10 @@
+.problemLayout {
+    width:90vw;
+    margin: 0 auto;
+    .problemLayoutBody {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+}
+
